numeric: Somewhat fix quiets
It's messy because we don't have extended bans (yet?) and +q is considered owner. For now let's just leave
it how it is and we'll move quiets to an extended ban or something. Coincidently, capture replies collide
with quiet list replies. So let's move capture replies somewhere else because they aren't really a "standard".
IRC is a joy.